fieldArray: Decoratorize
