+++ packages/redux-forms/src/index.ts
@@ -1,1 +1,7 @@
-export { default } from './formsReducer';
+import reducer, { State } from './formsReducer';
+
+export default reducer;
+
+export interface IReduxFormsState {
+  reduxForms: State;
+}
